chore(mypy): 量子化/NAS スクリプト7本を strict 化 allowlist 29->36
py.typed 後の unused-ignore 一掃を継続 + 軽微な実型修正:
- int8_quant_footprint/gptq_compare/quant_bitwidth_sweep/quant_group_compare/qat_train:
  llcore.lm.* import の不要 type:ignore[import-untyped] 除去(純機械)。
- nas_level2/evolve_linearization: tokenizer 行の不要 type:ignore[operator] 除去 +
  fitness 戻り値の no-any-return を float() / f:float 注釈で解消。
複合 mypy src+allowlist(36本) --strict = 103ファイル 0、ruff clean、関連79テスト green。

+++ scripts/nas_level2.py
@@ -88,7 +88,7 @@ def main(argv: list[str] | None = None) -> int:
         )
 
     text = Path(args.text_file).read_text(encoding="utf-8")[50000:]
-    ids = tok(text, return_tensors="pt").input_ids[:, : args.n_tokens]  # type: ignore[operator]
+    ids = tok(text, return_tensors="pt").input_ids[:, : args.n_tokens]
     base = nll(ids)
     print(f"[base] all-softmax nll={base:.4f} ppl={torch.tensor(base).exp():.2f}  mem/layer={mem_softmax}B @ctx{Tref}", flush=True)
 
@@ -110,7 +110,7 @@ def main(argv: list[str] | None = None) -> int:
 
     def fitness(genome: CatGenome) -> float:
         pct, dn = evaluate(genome)
-        return pct - args.penalty * 100.0 * max(0.0, dn - args.budget)
+        return float(pct - args.penalty * 100.0 * max(0.0, dn - args.budget))
 
     # budget-greedy baseline: per layer (tolerance order) pick the cheapest mixer that keeps
     # cumulative Δnll <= budget (try linear, then sliding, then softmax).
